refactor: extract useClassInfo
